seperationg suggestion logic
