+++ src/app/classes/suggestion-user-interface.spec.ts
@@ -1,0 +1,7 @@
+import { SuggestionUserInterface } from './suggestion-user-interface';
+
+describe('SuggestionUserInterface', () => {
+  it('should create an instance', () => {
+    expect(new SuggestionUserInterface(null)).toBeTruthy();
+  });
+});
